requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
aiogram==3.4.1
aiohttp==3.9.1
python-dotenv==1.0.0
requests==2.31.0
pytz==2024.1

